server/requirements.txt
asgiref==3.11.1
awsebcli==3.27.1
bcrypt==5.0.0
blessed==1.27.0
botocore==1.42.78
cement==2.10.14
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
colorama==0.4.6
cryptography==46.0.6
decorator==5.2.1
Deprecated==1.3.1
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
fabric==3.2.2
idna==3.11
invoke==2.2.1
jmespath==1.1.0
packaging==26.0
paramiko==4.0.0
pathspec==0.12.1
psycopg2-binary==2.9.11
pycparser==3.0
PyNaCl==1.6.2
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.3
requests==2.33.0
semantic-version==2.10.0
six==1.17.0
sqlparse==0.5.5
termcolor==2.5.0
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.2.14
wrapt==2.1.2
